Updated README and initial Documentation. Provided basic app samples.
Closes #30
Closes #40
Closes #26
